Move message.js to errors MU-17
